@@ -3,7 +3,7 @@ name = "argilla-server"
 dynamic = ["version"]
 description = "Open-source tool for exploring, labeling, and monitoring data for NLP projects."
 readme = "README.md"
-requires-python = ">=3.8,<3.11"
+requires-python = ">=3.9"
 license = { text = "Apache-2.0" }
 keywords = [
     "data-science",
@@ -34,8 +34,6 @@ dependencies = [
     "greenlet ~= 3.1.0",
     # Async SQLite
     "aiosqlite == 0.20.0",
-    # metrics
-    "scikit-learn >= 0.24.2",
     # Statics server
     "aiofiles >= 0.6,< 22.2",
     "PyYAML >= 5.4.1,< 6.1.0",
@@ -51,8 +49,6 @@ dependencies = [
     "rq ~= 1.16.2",
     # Info status
     "psutil >= 5.8, <5.10",
-    # Telemetry
-    "segment-analytics-python == 2.2.0",
     # For logging, tracebacks, printing, progressbars
     "rich != 13.1.0",
     # For CLI
@@ -84,6 +80,7 @@ build-backend = "pdm.backend"
 
 [tool.pdm]
 distribution = true
+ignore_package_warnings = true
 
 [tool.pdm.build]
 excludes = [".env.dev", ".env.test"]
@@ -102,9 +99,9 @@ test = [
     "pytest-env>=1.1.3",
     "factory-boy~=3.2.1",
     "httpx>=0.26.0",
-    # Required by tests/unit/utils/test_dependency.py but we should take a look a probably removed them
-    "spacy>=3.5.0,<3.7.0",
     "pytest-randomly>=3.15.0",
+    # pytest-randomly requires numpy < 2.0.0
+    "numpy<2.0.0",
 ]
 
 [tool.pytest.ini_options]
