@@ -12,61 +12,71 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from enum import Enum
+
+try:
+    from enum import StrEnum
+except ImportError:
+    from enum import Enum
+
+    class StrEnum(str, Enum):
+        """Custom StrEnum class for Python <3.11 compatibility."""
+
+        def __str__(self):
+            return str(self.value)
 
 
-class FieldType(str, Enum):
+class FieldType(StrEnum):
     text = "text"
     image = "image"
     chat = "chat"
     custom = "custom"
 
 
-class ResponseStatus(str, Enum):
+class ResponseStatus(StrEnum):
     draft = "draft"
     submitted = "submitted"
     discarded = "discarded"
 
 
-class ResponseStatusFilter(str, Enum):
+class ResponseStatusFilter(StrEnum):
     draft = "draft"
     pending = "pending"
     submitted = "submitted"
     discarded = "discarded"
 
 
-class SuggestionType(str, Enum):
+class SuggestionType(StrEnum):
     model = "model"
     human = "human"
 
 
-class DatasetStatus(str, Enum):
+class DatasetStatus(StrEnum):
     draft = "draft"
     ready = "ready"
 
 
-class DatasetDistributionStrategy(str, Enum):
+class DatasetDistributionStrategy(StrEnum):
     overlap = "overlap"
 
 
-class UserRole(str, Enum):
+class UserRole(StrEnum):
     owner = "owner"
     admin = "admin"
     annotator = "annotator"
 
 
-class RecordStatus(str, Enum):
+class RecordStatus(StrEnum):
     pending = "pending"
     completed = "completed"
 
 
-class RecordInclude(str, Enum):
+class RecordInclude(StrEnum):
     responses = "responses"
     suggestions = "suggestions"
     vectors = "vectors"
 
 
-class QuestionType(str, Enum):
+class QuestionType(StrEnum):
     text = "text"
     rating = "rating"
     label_selection = "label_selection"
@@ -75,13 +85,13 @@ class QuestionType(str, Enum):
     span = "span"
 
 
-class MetadataPropertyType(str, Enum):
+class MetadataPropertyType(StrEnum):
     terms = "terms"  # Textual types with a fixed value list
     integer = "integer"  # Integer values
     float = "float"  # Decimal values
 
 
-class RecordSortField(str, Enum):
+class RecordSortField(StrEnum):
     id = "id"
     external_id = "external_id"
     inserted_at = "inserted_at"
@@ -89,16 +99,16 @@ class RecordSortField(str, Enum):
     status = "status"
 
 
-class SortOrder(str, Enum):
+class SortOrder(StrEnum):
     asc = "asc"
     desc = "desc"
 
 
-class SimilarityOrder(str, Enum):
+class SimilarityOrder(StrEnum):
     most_similar = "most_similar"
     least_similar = "least_similar"
 
 
-class OptionsOrder(str, Enum):
+class OptionsOrder(StrEnum):
     natural = "natural"
     suggestion = "suggestion"
